moved Siva's highlightText to utils
