+++ ui/src/components/PipelineEditor/StepChooser.jsx
@@ -10,6 +10,7 @@ import { Spinner } from "../Spinner";
 import * as BonInABoxScriptService from "bon_in_a_box_script_service";
 import SubdirectoryArrowRightIcon from '@mui/icons-material/SubdirectoryArrowRight';
 import { Alert } from '@mui/material';
+import { highlightText } from "../../utils/HighlightText.jsx";
 
 const api = new BonInABoxScriptService.DefaultApi();
 
@@ -18,48 +19,6 @@ const onDragStart = (event, nodeType, descriptionFile) => {
   event.dataTransfer.setData("descriptionFile", descriptionFile);
   event.dataTransfer.effectAllowed = "move";
 };
-
-// Helper function to highlight text with search keywords
-function highlightText(text, searchQuery) {
-  if (!text || !searchQuery || searchQuery.trim() === "") {
-    return text;
-  }
-
-  const keywords = searchQuery.trim().split(/\s+/).filter(k => k.length > 0);
-  if (keywords.length === 0) {
-    return text;
-  }
-
-  // a case-insensitive regex that matches all keywords
-  const regexPattern = keywords
-    .map(keyword => keyword.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'))
-    .join('|');
-
-  const regex = new RegExp(`(${regexPattern})`, 'gi');
-  const parts = [];
-  let lastIndex = 0;
-  let keyCounter = 0;
-
-  // Find all matches and build JSX with highlights
-  const matches = Array.from(text.matchAll(regex));
-
-  for (const match of matches) {
-    // Add text before match
-    if (match.index > lastIndex) {
-      parts.push(text.substring(lastIndex, match.index));
-    }
-    // Add highlighted match
-    parts.push(<mark key={`highlight-${keyCounter++}`} className="search-highlight">{match[0]}</mark>);
-    lastIndex = match.index + match[0].length;
-  }
-
-  // Add remaining text
-  if (lastIndex < text.length) {
-    parts.push(text.substring(lastIndex));
-  }
-
-  return parts;
-}
 
 // Helper function to get metadata excerpt with highlighted keywords
 function getMetadataExcerpt(metadata, searchQuery) {
